test(doc): extract test_utils mod
