+++ crates/rari-doc/src/html/fix_img.rs
@@ -163,80 +163,11 @@ pub fn img_size(
 
 #[cfg(test)]
 mod tests {
-    use std::path::{Path, PathBuf};
-
     use lol_html::{RewriteStrSettings, element, rewrite_str};
-    use rari_types::RariEnv;
-    use rari_types::fm_types::{FeatureStatus, PageType};
-    use rari_types::locale::Locale;
     use url::Url;
 
     use super::handle_img;
-    use crate::error::DocError;
-    use crate::pages::page::PageLike;
-    use crate::pages::types::utils::FmTempl;
-
-    struct TestPage {
-        path: PathBuf,
-        url: String,
-    }
-
-    impl PageLike for TestPage {
-        fn url(&self) -> &str {
-            &self.url
-        }
-        fn slug(&self) -> &str {
-            ""
-        }
-        fn title(&self) -> &str {
-            ""
-        }
-        fn short_title(&self) -> Option<&str> {
-            None
-        }
-        fn locale(&self) -> Locale {
-            Locale::EnUs
-        }
-        fn content(&self) -> &str {
-            ""
-        }
-        fn rari_env(&self) -> Option<RariEnv<'_>> {
-            None
-        }
-        fn render(&self) -> Result<String, DocError> {
-            Ok(String::new())
-        }
-        fn title_suffix(&self) -> Option<&str> {
-            None
-        }
-        fn page_type(&self) -> PageType {
-            PageType::None
-        }
-        fn status(&self) -> &[FeatureStatus] {
-            &[]
-        }
-        fn full_path(&self) -> &Path {
-            &self.path
-        }
-        fn path(&self) -> &Path {
-            &self.path
-        }
-        fn base_slug(&self) -> &str {
-            ""
-        }
-        fn trailing_slash(&self) -> bool {
-            false
-        }
-        fn fm_offset(&self) -> usize {
-            0
-        }
-        fn raw_content(&self) -> &str {
-            ""
-        }
-        fn banners(&self) -> Option<&[FmTempl]> {
-            None
-        }
-    }
+    use crate::test_utils::TestPage;
 
     // Minimal GIF recognised by imagesize: the format detector reads a 12-byte
     // header, then the GIF size parser seeks to byte 6 and reads width/height
@@ -278,6 +209,7 @@ mod tests {
         let page = TestPage {
             path: tmp.join("index.md"),
             url: "/en-US/docs/Test".to_string(),
+            ..Default::default()
         };
 
         // comrak encodes `é` (U+00E9, UTF-8 0xC3 0xA9) as `%C3%A9`.
